LLgen: use size_t, reduce warnings, other small changes
Use C89 size_t for sizes from sizeof() or to malloc() or realloc().
Remove obsolete (unsigned) casts.  Sizes were unsigned int in
traditional C but are size_t in C89.

Silence some clang warnings.  Add the second pair of round brackets in
`while ((ff = ff->ff_next))` to silence -Wparentheses.  Change
`if (nc_first(...))/*nothing*/;` to `(void)nc_first(...);` to silence
-Wempty-body.  The code in compute.c nc_first() had the form
`if (x) if (y) s; else t;`.  The old indentation (before 10717cc)
suggests that the "else" belongs to the 2nd "if", so add braces like
`if (x) { if (y) s; else t; }` to silence -Wdangling-else.

Shuffle extern function declarations.  Add missing declaration for
LLparse().  Stop declaring RENAME(); it doesn't exist.  Move some
declarations from main.c to extern.h, so the C compiler may check that
the declarations are compatible with the function definitions.

Assume that standard C89 remove() is available and doesn't need the
UNLINK() wrapper.

In lib/incl, don't need to include <stdio.h> nor <stdlib.h> to use
assert().

Remove alloc.h.  If you don't clean your build, then an outdated
BUILDDIR/obj/util/LLgen/headers/alloc.h will survive but should not
cause harm, because nothing includes it.  Don't need to remove alloc.h
from util/LLgen/distr.sh, because it isn't there.

Run the bootstrap to rebuild LLgen.c, Lpars.c, tokens.c.

+++ util/LLgen/src/gencode.c
@@ -20,7 +20,6 @@
 
 #include <stdlib.h>
 #include <stdio.h>
-# include "alloc.h"
 # include "types.h"
 # include "io.h"
 # include "extern.h"
@@ -102,7 +101,7 @@ STATIC void doclose(FILE *f)
 STATIC int *mk_tokenlist(void)
 {
 	register int i = ntokens;
-	register int *p = (int *) alloc((unsigned) (i * sizeof(int))) + i;
+	register int *p = (int *) alloc(i * sizeof(int)) + i;
 
 	while (i--)
 		*--p = -1;
@@ -295,7 +294,7 @@ STATIC void genrecovery(void)
 	for (psetl = setptr; psetl < maxptr; psetl++)
 		prset(*psetl);
 	fputs(c_arrend, f);
-	index = (int *) alloc((unsigned) (assval * sizeof(int)));
+	index = (int *) alloc(assval * sizeof(int));
 	for (q = index; q < &index[assval];)
 		*q++ = -1;
 	for (t = tokens; t < maxt; t++)
@@ -344,7 +343,7 @@ STATIC void genncrecovery(void)
 	fprintf(f, "#define LLFIRST_NT %d\n", assval);
 	fprintf(f, "#define LLSETSIZE %d\n", nbytes);
 
-	index = (int *) alloc((unsigned) (assval * sizeof(int)));
+	index = (int *) alloc(assval * sizeof(int));
 	for (q = index; q < &index[assval];) *q++ = -1;
 	for (t = tokens; t < maxt; t++)
 	{
@@ -988,14 +987,14 @@ STATIC int *dopush(register p_gram p, int safety, int toplevel, int **pp)
 	/*
 	 * The safety only matters if toplevel != 0
 	 */
-	unsigned int i = 100;
+	size_t i = 100;
 	register int *ip = (int *) alloc(100 * sizeof(int));
 
 	*pp = ip;
 
 	for (;;)
 	{
-		if (ip - *pp >= i)
+		if ((size_t)(ip - *pp) >= i)
 		{
 			*pp = (int *) ralloc((p_mem) (*pp), (i + 100) * sizeof(int));
 			ip = *pp + i;
